requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pytesseract==0.3.10
Pillow==10.0.0
Flask==3.0.2